requirementsPython.txt
jpype1
pyTelegramBotAPI
Crypto
